Rilascio versione 3.8.0

+++ jars/orm-beans/src/main/java/it/govpay/model/TipoVersamento.java
@@ -19,7 +19,7 @@
  */
 package it.govpay.model;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public class TipoVersamento extends BasicModel {
 	private static final long serialVersionUID = 1L;
@@ -69,7 +69,7 @@ public class TipoVersamento extends BasicModel {
 	
 	private boolean avvisaturaMailPromemoriaScadenzaAbilitatoDefault;
 	private String avvisaturaMailPromemoriaScadenzaTipoDefault;
-	private BigDecimal avvisaturaMailPromemoriaScadenzaPreavvisoDefault;
+	private BigInteger avvisaturaMailPromemoriaScadenzaPreavvisoDefault;
 	private String avvisaturaMailPromemoriaScadenzaOggettoDefault;
 	private String avvisaturaMailPromemoriaScadenzaMessaggioDefault;
 	
@@ -99,7 +99,7 @@ public class TipoVersamento extends BasicModel {
 	
 	private boolean avvisaturaAppIoPromemoriaScadenzaAbilitatoDefault;
 	private String avvisaturaAppIoPromemoriaScadenzaTipoDefault;
-	private BigDecimal avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault;
+	private BigInteger avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault;
 	private String avvisaturaAppIoPromemoriaScadenzaOggettoDefault;
 	private String avvisaturaAppIoPromemoriaScadenzaMessaggioDefault;
 	
@@ -309,10 +309,10 @@ public class TipoVersamento extends BasicModel {
 	public void setAvvisaturaMailPromemoriaScadenzaTipoDefault(String avvisaturaMailPromemoriaScadenzaTipoDefault) {
 		this.avvisaturaMailPromemoriaScadenzaTipoDefault = avvisaturaMailPromemoriaScadenzaTipoDefault;
 	}
-	public BigDecimal getAvvisaturaMailPromemoriaScadenzaPreavvisoDefault() {
+	public BigInteger getAvvisaturaMailPromemoriaScadenzaPreavvisoDefault() {
 		return avvisaturaMailPromemoriaScadenzaPreavvisoDefault;
 	}
-	public void setAvvisaturaMailPromemoriaScadenzaPreavvisoDefault(BigDecimal avvisaturaMailPromemoriaScadenzaPreavvisoDefault) {
+	public void setAvvisaturaMailPromemoriaScadenzaPreavvisoDefault(BigInteger avvisaturaMailPromemoriaScadenzaPreavvisoDefault) {
 		this.avvisaturaMailPromemoriaScadenzaPreavvisoDefault = avvisaturaMailPromemoriaScadenzaPreavvisoDefault;
 	}
 	public String getAvvisaturaMailPromemoriaScadenzaOggettoDefault() {
@@ -424,10 +424,10 @@ public class TipoVersamento extends BasicModel {
 	public void setAvvisaturaAppIoPromemoriaScadenzaTipoDefault(String avvisaturaAppIoPromemoriaScadenzaTipoDefault) {
 		this.avvisaturaAppIoPromemoriaScadenzaTipoDefault = avvisaturaAppIoPromemoriaScadenzaTipoDefault;
 	}
-	public BigDecimal getAvvisaturaAppIoPromemoriaScadenzaPreavvisoDefault() {
+	public BigInteger getAvvisaturaAppIoPromemoriaScadenzaPreavvisoDefault() {
 		return avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault;
 	}
-	public void setAvvisaturaAppIoPromemoriaScadenzaPreavvisoDefault(BigDecimal avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault) {
+	public void setAvvisaturaAppIoPromemoriaScadenzaPreavvisoDefault(BigInteger avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault) {
 		this.avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault = avvisaturaAppIoPromemoriaScadenzaPreavvisoDefault;
 	}
 	public String getAvvisaturaAppIoPromemoriaScadenzaOggettoDefault() {
