+++ jars/orm/src/main/java/it/govpay/bd/model/TipoVersamentoDominio.java
@@ -19,7 +19,7 @@
  */
 package it.govpay.bd.model;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.openspcoop2.generic_project.exception.NotFoundException;
 import org.openspcoop2.generic_project.exception.ServiceException;
@@ -65,6 +65,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 			try {
 				this.tipoVersamento = AnagraficaManager.getTipoVersamento(configWrapper, this.getIdTipoVersamento());
 			} catch (NotFoundException e) {
+				// donothing
 			}
 		} 
 		return this.tipoVersamento;
@@ -75,6 +76,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 			try {
 				this.dominio = AnagraficaManager.getDominio(configWrapper, this.getIdDominio());
 			} catch (NotFoundException e) {
+				// donothing
 			}
 		} 
 		return this.dominio;
@@ -258,7 +260,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 		else 
 			return this.getAvvisaturaMailPromemoriaScadenzaTipoDefault();
 	}
-	public BigDecimal getAvvisaturaMailPromemoriaScadenzaPreavviso() {
+	public BigInteger getAvvisaturaMailPromemoriaScadenzaPreavviso() {
 		if(this.getAvvisaturaMailPromemoriaScadenzaPreavvisoCustom() != null)
 			return this.getAvvisaturaMailPromemoriaScadenzaPreavvisoCustom();
 		else 
@@ -385,7 +387,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 		else 
 			return this.getAvvisaturaAppIoPromemoriaScadenzaTipoDefault();
 	}
-	public BigDecimal getAvvisaturaAppIoPromemoriaScadenzaPreavviso() {
+	public BigInteger getAvvisaturaAppIoPromemoriaScadenzaPreavviso() {
 		if(this.getAvvisaturaAppIoPromemoriaScadenzaPreavvisoCustom() != null)
 			return this.getAvvisaturaAppIoPromemoriaScadenzaPreavvisoCustom();
 		else
